feature: Adds engine presentation separation
