+++ packages/engine/src/plan/status.test.ts
@@ -1,84 +1,9 @@
 import { describe, it, expect } from "vitest";
 import { parsePlan } from "./schema.js";
-import { getPlanStatus, formatDefaultView, formatFullView } from "./status.js";
-import type { DeviationReport } from "./detect.js";
+import { getPlanStatus } from "./status.js";
+import { makeFullPlan, makeSingleFractalPlan } from "./status.fixtures.js";
 
-// Fixed date for deterministic tests
 const TODAY = "2026-07-10";
-
-function makeSingleFractalPlan(weeks: object[], extras?: object) {
-  return parsePlan({
-    schema_version: 1,
-    block: "build",
-    start: "2026-05-04",
-    ...extras,
-    mesocycles: [{ name: "CANAL", fractals: [{ weeks }] }],
-  });
-}
-
-/**
- * Multi-mesocycle plan used for structural and full-view tests.
- * Week sequence (absolute index):
- *   CANAL / F1: L(1)ok  LL(2)ok  LLL(3)ok  D(4)ok  Ta(5)ok  Tb(6)ok
- *   CANAL / F2: L(7)current  LL(8)unsynced_past  LLL(9)unsynced_past
- *               D(10)future  Ta(11)future  Tb(12)future
- *   TAPER / F1: P(13).  P(14).  R(15).  N(16).
- *
- * TODAY = "2026-07-10"
- * week end dates after current:
- *   LL  starts 2026-06-22, end 2026-06-29 → past   → unsynced_past
- *   LLL starts 2026-06-29, end 2026-07-06 → past   → unsynced_past
- *   D   starts 2026-07-06, end 2026-07-13 → >=TODAY → future
- */
-function makeFullPlan() {
-  return parsePlan({
-    schema_version: 1,
-    block: "build",
-    goal: "Half Marathon Santiago",
-    race_date: "2026-10-18",
-    start: "2026-05-04",
-    mesocycles: [
-      {
-        name: "CANAL",
-        fractals: [
-          {
-            weeks: [
-              { planned: "L", start: "2026-05-04", executed: "L" },
-              { planned: "LL", start: "2026-05-11", executed: "LL" },
-              { planned: "LLL", start: "2026-05-18", executed: "LLL" },
-              { planned: "D", start: "2026-05-25", executed: "D" },
-              { planned: "Ta", start: "2026-06-01", executed: "Ta" },
-              { planned: "Tb", start: "2026-06-08", executed: "Tb" },
-            ],
-          },
-          {
-            weeks: [
-              { planned: "L", start: "2026-06-15" },   // 7 — current
-              { planned: "LL", start: "2026-06-22" },  // 8 — unsynced_past
-              { planned: "LLL", start: "2026-06-29" }, // 9 — unsynced_past
-              { planned: "D", start: "2026-07-06" },   // 10 — future
-              { planned: "Ta", start: "2026-07-13" },  // 11 — future
-              { planned: "Tb", start: "2026-07-20" },  // 12 — future
-            ],
-          },
-        ],
-      },
-      {
-        name: "TAPER",
-        fractals: [
-          {
-            weeks: [
-              { planned: "P", start: "2026-10-04" },  // 13 — future
-              { planned: "P", start: "2026-10-11" },  // 14 — future
-              { planned: "R", start: "2026-10-18" },  // 15 — future
-              { planned: "N", start: "2026-10-25" },  // 16 — future
-            ],
-          },
-        ],
-      },
-    ],
-  });
-}
 
 // ---------------------------------------------------------------------------
 // Engine computation tests
@@ -272,251 +197,5 @@ describe("getPlanStatus", () => {
     expect(ok).toHaveLength(2);
     expect(deviated).toHaveLength(1);
     expect(deviated[0]?.reason).toBe("travel");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Formatting tests
-// ---------------------------------------------------------------------------
-
-describe("formatDefaultView", () => {
-  it("formats default view with header, position, and milestones", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("BUILD — Half Marathon Santiago (2026-10-18)");
-    expect(output).toContain("Mesocycle: CANAL | Fractal 2 of 2");
-    expect(output).toContain("Week 7/16 — L (2026-06-15)");
-    expect(output).toContain("Next:");
-  });
-
-  it("shows unsynced past weeks section when present", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("Unsynced:");
-    expect(output).toContain("Week 8/16 — LL (2026-06-22)");
-    expect(output).toContain("Week 9/16 — LLL (2026-06-29)");
-  });
-
-  it("shows plan complete message when all weeks have executed", () => {
-    const plan = makeSingleFractalPlan([
-      { planned: "L", start: "2026-05-04", executed: "L" },
-      { planned: "LL", start: "2026-05-11", executed: "LL" },
-    ]);
-    const status = getPlanStatus(plan, "2026-06-01");
-    const output = formatDefaultView(status);
-    expect(output).toContain("Plan complete");
-    expect(output).toContain("2");
-  });
-});
-
-describe("formatFullView", () => {
-  it("formats full view with all mesocycles and fractal rows", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-
-    expect(output).toContain("BUILD — Half Marathon Santiago (2026-10-18)");
-    expect(output).toContain("CANAL");
-    expect(output).toContain("TAPER");
-    expect(output).toContain("F1:");
-    expect(output).toContain("F2:");
-  });
-
-  it("full view shows current week marker", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-
-    expect(output).toContain("^ current");
-  });
-
-  it("formats ok weeks as 'planned ok'", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-
-    expect(output).toContain("L ok");
-    expect(output).toContain("LL ok");
-  });
-
-  it("formats deviated weeks as 'executed[planned/reason]'", () => {
-    const plan = parsePlan({
-      schema_version: 1,
-      block: "build",
-      start: "2026-05-04",
-      mesocycles: [
-        {
-          name: "CANAL",
-          fractals: [
-            {
-              weeks: [
-                {
-                  planned: "LLL",
-                  start: "2026-05-04",
-                  executed: "INC",
-                  reason: "illness",
-                },
-                { planned: "D", start: "2026-05-11" }, // current
-              ],
-            },
-          ],
-        },
-      ],
-    });
-    const status = getPlanStatus(plan, "2026-05-14");
-    const output = formatFullView(status);
-    expect(output).toContain("INC[LLL/illness]");
-  });
-
-  it("formats unsynced past weeks with ? marker", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-    // LL and LLL in F2 are unsynced past
-    expect(output).toContain("LL?");
-    expect(output).toContain("LLL?");
-  });
-
-  it("formats future weeks with . marker", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-    expect(output).toContain("D .");
-    expect(output).toContain("P .");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Deviation surfacing tests (issue #18)
-// ---------------------------------------------------------------------------
-
-/** Minimal deviation report with no anomalies. */
-const CLEAN_REPORT: DeviationReport = {
-  suggestINC: false,
-  suggestDNF: false,
-  missingLongRunDay: undefined,
-  completedRuns: 5,
-  expectedRuns: 5,
-  incThreshold: 2,
-};
-
-/** Deviation report with INC suggestion and missing long run day. */
-const ANOMALOUS_REPORT: DeviationReport = {
-  suggestINC: true,
-  suggestDNF: false,
-  missingLongRunDay: "sunday",
-  completedRuns: 2,
-  expectedRuns: 5,
-  incThreshold: 2,
-};
-
-describe("deviation surfacing in default view", () => {
-  it("shows unsynced weeks with anomalies separately from clean unsynced weeks", () => {
-    const plan = makeFullPlan();
-    // week 8 (LL) has anomalies, week 9 (LLL) is clean
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, ANOMALOUS_REPORT],
-      [9, CLEAN_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("Unsynced with anomalies:");
-    expect(output).toContain("Week 8/16 — LL (2026-06-22)");
-    expect(output).toContain("Unsynced:");
-    expect(output).toContain("Week 9/16 — LLL (2026-06-29)");
-  });
-
-  it("shows anomaly detail line with run count and missing long run day", () => {
-    const plan = makeFullPlan();
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, ANOMALOUS_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("2/5 runs");
-    expect(output).toContain("missing long run day (sunday)");
-  });
-
-  it("omits anomalies section when no anomalous unsynced weeks", () => {
-    const plan = makeFullPlan();
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, CLEAN_REPORT],
-      [9, CLEAN_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatDefaultView(status);
-
-    expect(output).not.toContain("Unsynced with anomalies:");
-    expect(output).toContain("Unsynced:");
-  });
-
-  it("omits clean unsynced section when all unsynced weeks have anomalies", () => {
-    const plan = makeFullPlan();
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, ANOMALOUS_REPORT],
-      [9, ANOMALOUS_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("Unsynced with anomalies:");
-    // "Unsynced:" should not appear as a standalone section (without "anomalies")
-    // The full "Unsynced:" string also appears inside "Unsynced with anomalies:"
-    // so check it doesn't have the standalone group header
-    const lines = output.split("\n");
-    expect(lines.some((l) => l === "Unsynced:")).toBe(false);
-  });
-
-  it("shows plain Unsynced section when no deviation reports provided (backward compat)", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatDefaultView(status);
-
-    expect(output).toContain("Unsynced:");
-    expect(output).not.toContain("Unsynced with anomalies:");
-  });
-});
-
-describe("deviation surfacing in full view (?/??)", () => {
-  it("uses ?? marker for unsynced past weeks with anomalies", () => {
-    const plan = makeFullPlan();
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, ANOMALOUS_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatFullView(status);
-
-    expect(output).toContain("LL??");
-  });
-
-  it("uses ? marker for unsynced past weeks without anomalies", () => {
-    const plan = makeFullPlan();
-    const deviationReports = new Map<number, DeviationReport>([
-      [8, ANOMALOUS_REPORT],
-      [9, CLEAN_REPORT],
-    ]);
-    const status = getPlanStatus(plan, TODAY, { deviationReports });
-    const output = formatFullView(status);
-
-    expect(output).toContain("LL??");
-    expect(output).toContain("LLL?");
-    // LLL? should not be LLL??
-    expect(output).not.toContain("LLL??");
-  });
-
-  it("uses ? for all unsynced past weeks when no deviation reports provided", () => {
-    const plan = makeFullPlan();
-    const status = getPlanStatus(plan, TODAY);
-    const output = formatFullView(status);
-
-    expect(output).toContain("LL?");
-    expect(output).toContain("LLL?");
-    expect(output).not.toContain("??");
   });
 });
