+++ packages/engine/src/plan/status.ts
@@ -1,6 +1,5 @@
 import type { Plan } from "./types.js";
 import type { DeviationReport } from "./detect.js";
-import { reportHasAnomalies } from "./detect.js";
 import { walkPlan } from "./walk.js";
 
 // ---------------------------------------------------------------------------
@@ -170,169 +169,4 @@ export function getPlanStatus(
     unsyncedPastWeeks,
     weeks,
   };
-}
-
-// ---------------------------------------------------------------------------
-// Formatting helpers (shared between default and full views)
-// ---------------------------------------------------------------------------
-
-function buildHeader(status: PlanStatus): string {
-  let header = status.block.toUpperCase();
-  if (status.goal) {
-    header += ` — ${status.goal}`;
-    if (status.raceDate) {
-      header += ` (${status.raceDate})`;
-    }
-  }
-  return header;
-}
-
-function relativeLabel(weeksFromNow: number): string {
-  return weeksFromNow === 1 ? "next week" : `in ${weeksFromNow} weeks`;
-}
-
-/** Token for a week entry in the full structural view. */
-function weekToFullToken(w: WeekStatusEntry): string {
-  switch (w.marker) {
-    case "ok":
-      return `${w.planned} ok`;
-    case "deviated": {
-      const suffix = w.reason ? `/${w.reason}` : "";
-      return `${w.executed}[${w.planned}${suffix}]`;
-    }
-    case "current":
-      // No suffix — distinguished by the `^ current` line below
-      return w.planned;
-    case "unsynced_past": {
-      const marker = w.deviationReport && reportHasAnomalies(w.deviationReport) ? "??" : "?";
-      return `${w.planned}${marker}`;
-    }
-    case "future":
-      return `${w.planned} .`;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Default view
-// ---------------------------------------------------------------------------
-
-/**
- * Formats the default (current-week-focused) plan status view.
- */
-export function formatDefaultView(status: PlanStatus): string {
-  const lines: string[] = [];
-
-  lines.push(buildHeader(status));
-
-  if (status.isComplete) {
-    lines.push(`Plan complete. All ${status.totalWeeks} weeks executed.`);
-    return lines.join("\n");
-  }
-
-  const cw = status.currentWeek!;
-
-  lines.push(`Mesocycle: ${cw.mesocycleName} | Fractal ${cw.fractalIndex} of ${cw.totalFractals}`);
-  lines.push("");
-  lines.push(`Week ${cw.absoluteIndex}/${cw.totalWeeks} — ${cw.planned} (${cw.start})`);
-
-  if (status.nextMilestones.length > 0) {
-    const [first, second] = status.nextMilestones;
-    if (second) {
-      lines.push(
-        `  Next: ${first!.planned} (${relativeLabel(first!.weeksFromNow)}) → ${second.planned} in ${second.weeksFromNow} weeks`,
-      );
-    } else {
-      lines.push(`  Next: ${first!.planned} (${relativeLabel(first!.weeksFromNow)})`);
-    }
-  }
-
-  // Split unsynced past weeks into two groups: anomalous vs clean
-  const withAnomalies = status.unsyncedPastWeeks.filter(
-    (w) => w.deviationReport && reportHasAnomalies(w.deviationReport),
-  );
-  const cleanUnsynced = status.unsyncedPastWeeks.filter(
-    (w) => !w.deviationReport || !reportHasAnomalies(w.deviationReport),
-  );
-
-  if (withAnomalies.length > 0) {
-    lines.push("");
-    lines.push("Unsynced with anomalies:");
-    for (const w of withAnomalies) {
-      const r = w.deviationReport!;
-      const details: string[] = [];
-      details.push(`${r.completedRuns}/${r.expectedRuns} runs`);
-      if (r.missingLongRunDay) {
-        details.push(`missing long run day (${r.missingLongRunDay})`);
-      }
-      lines.push(
-        `  Week ${w.absoluteIndex}/${w.totalWeeks} — ${w.planned} (${w.start}): ${details.join(", ")}`,
-      );
-    }
-  }
-
-  if (cleanUnsynced.length > 0) {
-    lines.push("");
-    lines.push("Unsynced:");
-    for (const w of cleanUnsynced) {
-      lines.push(`  Week ${w.absoluteIndex}/${w.totalWeeks} — ${w.planned} (${w.start})`);
-    }
-  }
-
-  return lines.join("\n");
-}
-
-// ---------------------------------------------------------------------------
-// Full view
-// ---------------------------------------------------------------------------
-
-/**
- * Formats the full structural overview of the plan.
- * Each mesocycle is shown with its fractals; each week has a status marker.
- * The current week is identified with a `^ current` caret on the line below.
- */
-export function formatFullView(status: PlanStatus): string {
-  const lines: string[] = [];
-
-  lines.push(buildHeader(status));
-
-  // Group weeks into mesocycle → fractal buckets preserving plan order
-  interface MesoGroup {
-    name: string;
-    fractals: WeekStatusEntry[][];
-  }
-  const mesoGroups: MesoGroup[] = [];
-
-  for (const w of status.weeks) {
-    let meso = mesoGroups.find((m) => m.name === w.mesocycleName);
-    if (!meso) {
-      meso = { name: w.mesocycleName, fractals: [] };
-      mesoGroups.push(meso);
-    }
-    const fi = w.fractalIndex - 1; // 0-based
-    while (meso.fractals.length <= fi) meso.fractals.push([]);
-    meso.fractals[fi]!.push(w);
-  }
-
-  for (const meso of mesoGroups) {
-    lines.push("");
-    lines.push(meso.name);
-
-    meso.fractals.forEach((fractalWeeks, fi) => {
-      const prefix = `  F${fi + 1}: `;
-      const tokens = fractalWeeks.map(weekToFullToken);
-      lines.push(prefix + tokens.join("  "));
-
-      const currentIdx = fractalWeeks.findIndex((w) => w.marker === "current");
-      if (currentIdx >= 0) {
-        // Compute character column of the current week token
-        let col = prefix.length;
-        for (let i = 0; i < currentIdx; i++) {
-          col += tokens[i]!.length + 2; // token + 2-space separator
-        }
-        lines.push(" ".repeat(col) + "^ current");
-      }
-    });
-  }
-
-  return lines.join("\n");
 }
